Insufficient timeout when calling FRR or BIRD over REST API for "ping"
With ping on FRR and BIRD using by default a count of 5 (-c 5) and no response timeout (-W), the resulting time it takes for a ping to an unavailable host is too short. As a result end-users see the message that "an error has occured" instead of "Unable to reach x.x.x.x.".
Increasing the API call timeout to 15 seconds fixes this.