Fix field validation for device config.
Some of the field validator methods for device config used `values` as a dictionary (presumably from an old version of pydantic). This is now passed as a `ValidationInfo` object by pydantic, which caused issues when code attempted to access it. The affected methods have been updated to fix the issue.

Fix #311

Signed-off-by: Mattie Nickson <mnickson@sidingsmedia.com>